requirements.txt
openai
sounddevice
numpy
scipy
pygame
python-dotenv
pydub
